Move EndlessRoadOracle into a 'dev/' directory, because I am told that then I can get it into master without enduring a review upfront.
