+++ drake/automotive/CMakeLists.txt
@@ -1,3 +1,4 @@
+add_subdirectory(dev)
 add_subdirectory(gen)
 add_subdirectory(maliput)
 
@@ -9,7 +10,7 @@ add_library_with_exports(LIB_NAME drakeAutomotive SOURCE_FILES
   create_trajectory_params.cc
   curve2.cc
   endless_road_car.cc
-  endless_road_oracle.cc
+  dev/endless_road_oracle.cc
   gen/driving_command.cc
   gen/endless_road_car_config.cc
   gen/endless_road_car_state.cc
@@ -50,7 +51,6 @@ drake_install_headers(
   create_trajectory_params.h
   curve2.h
   endless_road_car.h
-  endless_road_oracle.h
   idm_planner.h
   linear_car.h
   simple_car.h
