+++ drake/automotive/dev/CMakeLists.txt
@@ -1,0 +1,7 @@
+if(BUILD_TESTING)
+  add_subdirectory(test)
+endif()
+
+drake_install_headers(
+  endless_road_oracle.h
+  )
