+++ drake/automotive/dev/endless_road_oracle-internal.h
