+++ drake/automotive/dev/endless_road_oracle.h
