+++ drake/automotive/dev/test/CMakeLists.txt
@@ -1,0 +1,3 @@
+drake_add_cc_test(endless_road_oracle_test)
+target_link_libraries(endless_road_oracle_test
+                      drakeAutomotive)
